[TriggerExtractor] extract logic to TriggerMessageResolver
